requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
